requirements.txt
requests==2.32.3
python-dotenv==1.0.1
python-whois==0.9.4
fpdf2==2.8.1
colorama==0.4.6